pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "statmerge"
version = "1.0.0"
description = "Unified statistical analysis and agreement metrics with a desktop GUI"
authors = [{name="Mirza Niaz Zaman Elin"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "PySide6",
  "numpy",
  "pandas",
  "matplotlib",
  "scikit-learn",
  "python-docx"
]

[project.urls]
Homepage = "https://example.com/statmerge"
Issues = "https://example.com/statmerge/issues"

[project.scripts]
statmerge = "statmerge.cli:main"

[tool.pytest.ini_options]
addopts = "-q"
